move fn to BitBuffer
