+++ vortex-buffer/Cargo.toml
@@ -42,6 +42,7 @@ workspace = true
 [dev-dependencies]
 arrow-buffer = { workspace = true }
 divan = { workspace = true }
+rstest = { workspace = true }
 
 [[bench]]
 name = "vortex_buffer"
